Sprite rendering bug fix Sprite position should be stored as signed integers, as the value in memory is 8 or 16 bits higher than the real position. Therefore when subtracting, an unsigned integer would instead underflow into a high value, and even adding up the current x later won't help as it will be automatically promoted to a higher capacity integer.

+++ GBEmu/GBEmu/Include/GPU_Sprite.h
@@ -34,8 +34,8 @@ public:
 		Sprite::Size sprite_size, size_t current_line) const;
 
 private:
-	uint8_t y_position_{ 0 };
-	uint8_t x_position_{ 0 };
+	int16_t y_position_{ 0 };
+	int16_t x_position_{ 0 };
 	uint8_t tile_number_{ 0 };
 	bool is_above_background_{ true };
 	bool is_vertically_flipped_{ false };
